Rename and refactor Journey class
In order to more closely resemble the Trainline API, I've refactored the
server so that it can receive requests for the
`/locations-search/v2/search` endpoint. The `#search` method in the now
renamed JourneyApi class must call the `#location_code` method in order
to obtain the URNs for the "from" and "to" locations.