Handle errors thrown when the API is unavailable
As it is unclear how these errors should be handled, I've decided to
return an empty hash in the `ComThetrainline.find` method when the
server is down. Depending on decisions made within the team on how to
structure the response bodies, an explicit error message could be added
to this method in the future.